+++ CHANGELOG.md
@@ -14,6 +14,7 @@ request to fix it.
 ### Changed
 
 - [lustre/event] Correctly document the unit used for `debounce` and `throttle`.
+- [lustre/runtime] Fix a bug where holding a `WeakRef` to a context subscription function would cause it to GC unpredictably.
 - [lustre/server_component] Fixed a bug where listening to attribute changes would crash the JavaScript server component runtime.
 
 ## [v5.3.4] - 2025-08-23